# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"Service is healthy"}},"summary":"Health check endpoint","tags":["Health"]}},"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Registration successful"},"409":{"description":"Email already exists"}},"summary":"Register a new user and trigger OTP email","tags":["Authentication"]}},"/api/v1/auth/verify":{"post":{"operationId":"AuthController_verifyOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyOtpDto"}}}},"responses":{"200":{"description":"Email verified, tokens returned"},"400":{"description":"Invalid or expired OTP"}},"summary":"Verify OTP and activate account","tags":["Authentication"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Login successful, tokens returned"},"401":{"description":"Invalid credentials"}},"summary":"Login with email and password","tags":["Authentication"]}},"/api/v1/auth/resend-otp":{"post":{"operationId":"AuthController_resendOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendOtpDto"}}}},"responses":{"200":{"description":"OTP resent"}},"summary":"Resend OTP to email","tags":["Authentication"]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"New tokens returned"}},"summary":"Refresh access token using refresh token","tags":["Authentication"]}},"/api/v1/auth/me":{"post":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"User profile"}},"security":[{"bearer":[]}],"summary":"Get current authenticated user profile","tags":["Authentication"]}},"/api/v1/fx/rates":{"get":{"operationId":"FxRateController_getRates","parameters":[],"responses":{"200":{"description":"Returns all currency rates keyed to NGN"}},"security":[{"bearer":[]}],"summary":"Get current FX rates for all supported currencies","tags":["FX Rates"]}},"/api/v1/fx/rates/{currency}":{"get":{"operationId":"FxRateController_getRateByCurrency","parameters":[{"name":"currency","required":true,"in":"path","description":"Target currency code","schema":{"example":"USD","type":"string"}}],"responses":{"200":{"description":"Rate for the requested currency"}},"security":[{"bearer":[]}],"summary":"Get the FX rate for a specific currency pair against NGN","tags":["FX Rates"]}},"/api/v1/fx/supported":{"get":{"operationId":"FxRateController_getSupportedCurrencies","parameters":[],"responses":{"200":{"description":"Array of supported currency codes"}},"security":[{"bearer":[]}],"summary":"List all supported currency codes","tags":["FX Rates"]}},"/api/v1/wallet":{"get":{"operationId":"WalletController_getWallets","parameters":[],"responses":{"200":{"description":"List of wallets with balances"}},"security":[{"bearer":[]}],"summary":"Get all wallet balances for the authenticated user","tags":["Wallet"]}},"/api/v1/wallet/fund":{"post":{"operationId":"WalletController_fundWallet","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FundWalletDto"}}}},"responses":{"201":{"description":"Wallet funded successfully"},"400":{"description":"Validation error"}},"security":[{"bearer":[]}],"summary":"Fund wallet in NGN or another supported currency","tags":["Wallet"]}},"/api/v1/wallet/convert":{"post":{"operationId":"WalletController_convertCurrency","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConvertWalletDto"}}}},"responses":{"200":{"description":"Conversion successful"},"400":{"description":"Insufficient balance or invalid currencies"}},"security":[{"bearer":[]}],"summary":"Convert between currencies using real-time FX rates","tags":["Wallet"]}},"/api/v1/wallet/trade":{"post":{"operationId":"WalletController_trade","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeWalletDto"}}}},"responses":{"200":{"description":"Trade executed successfully"},"400":{"description":"Insufficient balance or invalid pair"}},"security":[{"bearer":[]}],"summary":"Trade NGN against another currency or vice versa","tags":["Wallet"]}},"/api/v1/transactions":{"get":{"operationId":"TransactionController_getTransactions","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"example":20,"type":"number"}},{"name":"type","required":false,"in":"query","schema":{"enum":["funding","conversion","trade","withdrawal"],"type":"string"}}],"responses":{"200":{"description":"Paginated transaction list"}},"security":[{"bearer":[]}],"summary":"Get paginated transaction history for current user","tags":["Transactions"]}},"/api/v1/transactions/{id}":{"get":{"operationId":"TransactionController_getTransaction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Transaction details"},"404":{"description":"Transaction not found"}},"security":[{"bearer":[]}],"summary":"Get a single transaction by ID (must belong to current user)","tags":["Transactions"]}}},"info":{"title":"FX Trading API","description":"Backend API for the FX Trading Application","version":"1.0.0","contact":{}},"tags":[{"name":"Authentication","description":"Register, verify, and login"},{"name":"Wallet","description":"Fund, convert, and trade currencies"},{"name":"FX Rates","description":"Real-time exchange rates"},{"name":"Transactions","description":"Transaction history and details"}],"servers":[],"components":{"securitySchemes":{"Bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"User email address","example":"user@example.com"},"password":{"type":"string","minLength":8,"maxLength":64,"description":"Account password (min 8 characters)","example":"SecurePass123!"}},"required":["email","password"]},"VerifyOtpDto":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"Registered email address","example":"user@example.com"},"otp":{"type":"string","minLength":6,"maxLength":6,"description":"6-digit OTP sent to your email","example":"123456"}},"required":["email","otp"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"Registered email address","example":"user@example.com"},"password":{"type":"string","description":"Account password","example":"SecurePass123!"}},"required":["email","password"]},"ResendOtpDto":{"type":"object","properties":{"email":{"type":"string","format":"email","description":"Email address to resend OTP to","example":"user@example.com"}},"required":["email"]},"FundWalletDto":{"type":"object","properties":{"amount":{"type":"number","minimum":1,"description":"Amount to fund","example":5000},"currency":{"type":"string","description":"Currency to fund (defaults to NGN)","example":"NGN"}},"required":["amount"]},"ConvertWalletDto":{"type":"object","properties":{"amount":{"type":"number","minimum":1,"description":"Amount to convert","example":1000},"fromCurrency":{"type":"string","description":"Currency to convert from","example":"NGN"},"toCurrency":{"type":"string","description":"Currency to convert to","example":"USD"}},"required":["amount","fromCurrency","toCurrency"]},"TradeWalletDto":{"type":"object","properties":{"amount":{"type":"number","minimum":1,"description":"Amount to trade","example":500},"fromCurrency":{"type":"string","description":"Currency you are selling","example":"NGN"},"toCurrency":{"type":"string","description":"Currency you want to buy","example":"USD"}},"required":["amount","fromCurrency","toCurrency"]}}}}